# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/griddlehq/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/2/abtests": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Lists experiments with Algolia-compatible pagination and optional index name\nprefix/suffix filtering.",
        "operationId": "list_experiments",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of experiments to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "indexPrefix",
            "in": "query",
            "description": "Filter experiments by index name prefix",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "indexSuffix",
            "in": "query",
            "description": "Filter experiments by index name suffix",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experiments list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgoliaListAbTestsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Experiment missing numeric ID mapping"
          },
          "503": {
            "description": "Experiment store unavailable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Creates a new experiment and returns the Algolia-compatible action payload\nusing the stable numeric experiment identifier.",
        "operationId": "create_experiment",
        "requestBody": {
          "description": "A/B test creation payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlgoliaCreateAbTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experiment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgoliaCreateAbTestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid experiment configuration"
          },
          "409": {
            "description": "Experiment already exists"
          },
          "500": {
            "description": "Experiment missing numeric ID mapping"
          },
          "503": {
            "description": "Experiment store unavailable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/2/abtests/estimate": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Estimate A/B test sample size and duration for a given minimum detectable effect and traffic split.",
        "description": "# Arguments\n\n* `State(state)` - Application state containing the optional analytics engine\n* `Json(body)` - A/B test request with variants and configuration including minimum detectable effect size\n\n# Returns\n\nAn HTTP response containing either an error response or an AlgoliaEstimateResponse with calculated sample sizes per variant and estimated duration in days.\n\nReturns an error if: variants count is not exactly 2, MDE size is outside (0, 1), or traffic percentages are not each in [1, 99] and do not sum to 100.\n\n# Behavior\n\nCalculates required sample size using a simplified formula with 95% confidence interval and 80% statistical power. Queries the analytics engine for historical daily search traffic (past 30 days) to estimate experiment duration. If traffic data is unavailable, uses a conservative default duration.",
        "operationId": "estimate_ab_test",
        "requestBody": {
          "description": "A/B test estimate payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlgoliaEstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Estimated sample sizes and duration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgoliaEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid estimate configuration"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/2/abtests/{id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Fetches a single experiment by numeric ID or UUID and enriches the payload\nwith analytics-backed results metadata when available.",
        "operationId": "get_experiment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Experiment identifier (numeric ID or UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgoliaAbTest"
                }
              }
            }
          },
          "404": {
            "description": "Experiment not found"
          },
          "500": {
            "description": "Experiment missing numeric ID mapping"
          },
          "503": {
            "description": "Experiment store unavailable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "experiments"
        ],
        "summary": "Replace an experiment's mutable configuration fields while preserving its\nidentity and lifecycle state.",
        "description": "Merges the request body with the existing experiment, keeping immutable\nidentity and lifecycle fields while falling back to existing values for\noptional settings omitted from the request.",
        "operationId": "update_experiment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Experiment identifier (numeric ID or UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated experiment configuration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experiment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgoliaAbTest"
                }
              }
            }
          },
          "400": {
            "description": "Invalid experiment configuration"
          },
          "404": {
            "description": "Experiment not found"
          },
          "409": {
            "description": "Experiment is in an invalid status for update"
          },
          "500": {
            "description": "Experiment missing numeric ID mapping"
          },
          "503": {
            "description": "Experiment store unavailable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "Deletes an experiment by numeric ID or UUID and returns the Algolia-style\naction envelope for the affected experiment.",
        "operationId": "delete_experiment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Experiment identifier (numeric ID or UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlgoliaAbTestActionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Experiment not found"
          },
          "409": {
            "description": "Cannot delete a running experiment"
          },
          "500": {
            "description": "Experiment missing numeric ID mapping"
          },
          "503": {
            "description": "Experiment store unavailable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/2/abtests/{id}/results": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Returns statistical results for an experiment addressed by numeric ID or UUID.",
        "operationId": "get_experiment_results",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Experiment identifier (numeric ID or UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment statistical results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Experiment not found"
          },
          "500": {
            "description": "Experiment missing numeric ID mapping"
          },
          "503": {
            "description": "Experiment store unavailable"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlgoliaCreateAbTestResponse": {
        "type": "object",
        "description": "Successful create response.",
        "required": [
          "abTestID",
          "index",
          "taskID"
        ],
        "properties": {
          "abTestID": {
            "type": "integer",
            "format": "int64"
          },
          "index": {
            "type": "string"
          },
          "taskID": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AlgoliaListAbTestsResponse": {
        "type": "object",
        "description": "List response for GET /2/abtests.",
        "required": [
          "count",
          "total"
        ],
        "properties": {
          "abtests": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/AlgoliaAbTest"
            }
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}